pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "eddy_bot"
description = "Python Discord Bot."
version = "0.2.0"
authors = [{ name = "Evan McKinney" }]
requires-python = ">=3.9"
dependencies = [
    "tenacity",
    "aiohttp",
    "aiosqlite",
    "discord.py",
    "numpy",
    "openai",
    "pyyaml",
    "pillow",
]

[project.optional-dependencies]
dev = ["ipykernel", "pre-commit", "pylatexenc"]
format = [
    "pre-commit",
    "black[jupyter]",
    "isort",
    "tomli",
    "pydocstyle",
    "ruff",
    "docformatter[tomli]",
]
test = ["pytest"]

[tool.ruff]
target-version = "py39"
fix = true

[tool.isort]
profile = "black"
